nsl/requirements.txt
absl==0.0
absl_py==0.12.0
catalyst==21.11
click==8.0.1
cutmix==0.1
dlib==19.22.1
matplotlib==3.2.2
mxnet==1.9.0
mxnet_cu102==1.6.0
numpy==1.21.5
op_cpp.egg==info
opencv_python==4.5.2.54
pandas==1.2.4
Pillow==9.0.0
pytorch_lightning==1.5.5
PyTurboJPEG==1.6.4
PyYAML==6.0
scipy==1.6.3
setuptools==47.1.0
tensorflow==2.5.2
tensorflow_addons==0.13.0
tensorflow_datasets==4.4.0
timm==0.4.12
torch==1.10.1+cu113
torchmetrics==0.5.1rc0
torchvision==0.11.2+cu113
tqdm==4.61.1
webdataset==0.1.62
